requirements.txt
httpx==0.25.0
python-dotenv==1.0.0
json-fix==0.5.2
pydantic==2.4.2
rich==13.6.0
